Get code to compile and run. Issues with inlined code in translateCmd.

+++ src/translateCmd.hpp
@@ -17,7 +17,7 @@ bool isExit(const std::string& inStr);
 //
 // Help command
 //
-bool isHelp(const std::string& inStr);
+bool isHelp(const std::string &inStr);
 
 //
 // Check if string indicates a valid move command
@@ -27,17 +27,17 @@ bool isMove(const std::string& inStr);
 //
 // Undo last move
 //
-bool isUndo(const std::string& inStr);
+bool isUndo(const std::string &inStr);
 
 //
 // Display board
 //
-bool isDisp(const std::string& inStr);
+bool isDisp(const std::string &inStr);
 
 //
 // Restart game without exiting
 //
-bool isRestart(const std::string& inStr);
+bool isRestart(const std::string &inStr);
 
 //
 // Help instructions
@@ -48,6 +48,6 @@ void displayHelp();
 // Translate a valid move command string to
 // a (row, col) pair
 //
-std::pair<uint8_t, uint8_t> cmd2Move(const std::string& inStr);
+std::pair<int, int> cmd2Move(const std::string& inStr);
 
 #endif
